requirements.txt
Flask==2.0.2
pycryptodome==3.15.0
